+++ src/main/java/org/realityforge/replicant/client/test/ReplicantClientTestModule.java
@@ -1,7 +1,5 @@
 package org.realityforge.replicant.client.test;
 
-import com.google.gwt.event.shared.EventBus;
-import com.google.gwt.event.shared.SimpleEventBus;
 import org.realityforge.guiceyloops.shared.AbstractModule;
 import org.realityforge.replicant.client.EntityChangeBroker;
 import org.realityforge.replicant.client.EntityChangeBrokerImpl;
@@ -19,7 +17,6 @@ public final class ReplicantClientTestModule
   @Override
   protected void configure()
   {
-    bindSingleton( EventBus.class, SimpleEventBus.class );
     bindSingleton( EntityRepository.class, EntityRepositoryImpl.class );
     bindSingleton( EntityChangeBroker.class, EntityChangeBrokerImpl.class );
     bindSingleton( EntitySubscriptionManager.class, EntitySubscriptionManagerImpl.class );