+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.5.57 (Pending):
+## 0.5.57:
 * Rename package `org.realityforge.replicant.client.json.gwt` to `org.realityforge.replicant.client.gwt`.
 * Extract the gwt specific functionality of AbstractClientTest to AbstractGwtClientTest and move
   to a separate directory.